feat: Enhance GGUF reader to support early v3 format and improve diagnostics in cross-validation

+++ crates/bitnet-models/src/formats/gguf/loader.rs
@@ -210,7 +210,7 @@ impl GgufLoader {
         device: &Device,
         config: &LoadConfig,
     ) -> Result<GgufTensors> {
-        let tensor_count = reader.tensor_count();
+        let tensor_count = reader.tensor_count() as usize;
         let mut tensors = GgufTensors::new();
 
         info!("Loading {} tensors", tensor_count);
